Simplify app removing soon-to-be deprecated features
- Transceive API
- microSD support
