+++ src/main/AndroidManifest.xml
@@ -14,23 +14,13 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name="com.fidesmo.sec.servicedeliverydemo.MainActivity"
+            android:name="com.fidesmo.sec.servicedeliverydemo.ServiceDeliveryActivity"
             android:label="@string/app_name"
-            android:screenOrientation="portrait" >
+            android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity
-            android:name="com.fidesmo.sec.servicedeliverydemo.TransceiveActivity"
-            android:label="@string/app_name"
-            android:screenOrientation="portrait">
-        </activity>
-        <activity
-            android:name="com.fidesmo.sec.servicedeliverydemo.ServiceDeliveryActivity"
-            android:label="@string/app_name"
-            android:screenOrientation="portrait">
         </activity>                
     </application>
