--- src/main/res/layout/activity_main.xml
@@ -1,37 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".MainActivity" >
-
-    <TextView
-        android:id="@+id/textView1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:textSize="18sp"
-        android:layout_marginBottom="20dp"
-        android:text="@string/introduction"
-        android:textAlignment="center" />
-
-    <Button
-        android:id="@+id/transceive_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/textView1"
-        android:text="@string/transceive_button" />
-
-    <Button
-        android:id="@+id/service_delivery_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/transceive_button"
-        android:layout_marginTop="32dp"
-        android:text="@string/service_delivery_button" />
-
-</RelativeLayout>
