+++ src/main/res/layout/activity_service_delivery.xml
@@ -70,13 +70,4 @@
         android:layout_marginTop="32dp"
         android:text="@string/card_button" />
 
-    <Button
-        android:id="@+id/devicefidelity_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/card_button"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="32dp"
-        android:text="@string/microsd_button" />
-
 </RelativeLayout>
